requirements.txt
fastapi == 0.116.1
uvicorn == 0.35.0
SQLAlchemy == 2.0.43
psutil == 7.0.0
APScheduler == 3.11.0
psycopg2-binary == 2.9.10
python-dotenv == 1.1.1
python-jose ==3.5.0
passlib == 1.7.4
python-multipart == 0.0.20
prometheus_client == 0.22.1
prometheus-fastapi-instrumentator == 7.1.0